refactor: 불필요한 파일 삭제 및 폴더 정리
- /pages/calendar/style 폴더 삭제
- calendar.style.tsx 삭제 후 css 는 calendar.tsx 에 합쳐서 작성
- app router 수정
